Refactor resolveSQLHandlebar
